add a script to update flutter plugins; rename files and classes to match the ones created by `flutter create`
