+++ flutter_libs/android/src/main/java/io/objectbox/objectbox_flutter_libs/ObjectboxFlutterLibsPlugin.java
@@ -1,4 +1,4 @@
-package io.objectbox.flutter;
+package io.objectbox.objectbox_flutter_libs;
 
 import androidx.annotation.NonNull;
 
@@ -8,8 +8,8 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 
-/** ObjectBoxFlutterPlugin */
-public class ObjectBoxFlutterPlugin implements FlutterPlugin, MethodCallHandler {
+/** ObjectboxFlutterLibsPlugin */
+public class ObjectboxFlutterLibsPlugin implements FlutterPlugin, MethodCallHandler {
   /// The MethodChannel that will the communication between Flutter and native Android
   ///
   /// This local reference serves to register the plugin with the Flutter Engine and unregister it
@@ -18,7 +18,7 @@ public class ObjectBoxFlutterPlugin implements FlutterPlugin, MethodCallHandler 
 
   @Override
   public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
-    channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "objectbox");
+    channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "objectbox_flutter_libs");
     channel.setMethodCallHandler(this);
   }
