--- flutter_libs/ios/Classes/ObjectBoxFlutterPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface ObjectBoxFlutterPlugin : NSObject<FlutterPlugin>
-@end
