+++ flutter_libs/ios/Classes/ObjectboxFlutterLibsPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface ObjectboxFlutterLibsPlugin : NSObject<FlutterPlugin>
+@end
