+++ flutter_libs/ios/Classes/SwiftObjectboxFlutterLibsPlugin.swift
@@ -1,10 +1,10 @@
 import Flutter
 import UIKit
 
-public class SwiftObjectBoxFlutterPlugin: NSObject, FlutterPlugin {
+public class SwiftObjectboxFlutterLibsPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "objectbox_sync_flutter_libs", binaryMessenger: registrar.messenger())
-    let instance = SwiftObjectBoxFlutterPlugin()
+    let channel = FlutterMethodChannel(name: "objectbox_flutter_libs", binaryMessenger: registrar.messenger())
+    let instance = SwiftObjectboxFlutterLibsPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
