+++ sync_flutter_libs/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 1d9032c7e1d867f071f2277eb1673e8f9b0274e3
+  channel: stable
+
+project_type: plugin
