+++ sync_flutter_libs/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="io.objectbox.flutter">
+  package="io.objectbox.objectbox_sync_flutter_libs">
 </manifest>
