+++ sync_flutter_libs/ios/Classes/ObjectboxSyncFlutterLibsPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface ObjectboxSyncFlutterLibsPlugin : NSObject<FlutterPlugin>
+@end
